add database schema trait
